+++ pioreactor/automations/temperature/__init__.py
@@ -2,8 +2,8 @@
 from __future__ import annotations
 
 from .constant_duty_cycle import ConstantDutyCycle
-from .only_record_ambient_temperature import OnlyRecordAmbientTemperature
+from .only_record_temperature import OnlyRecordTemperature
 from .thermostat import Thermostat
 
 
-__all__ = ("OnlyRecordAmbientTemperature", "ConstantDutyCycle", "Thermostat")
+__all__ = ("OnlyRecordTemperature", "ConstantDutyCycle", "Thermostat")