+++ pioreactor/background_jobs/temperature_control.py
@@ -228,14 +228,12 @@ class TemperatureController(BackgroundJob):
                 )
 
         averaged_temp = running_sum / running_count
-        if averaged_temp == 0.0 and self.automation_name != "only_record_ambient_temperature":
+        if averaged_temp == 0.0 and self.automation_name != "only_record_temperature":
             # this is a hardware fluke, not sure why, see #308. We will return something very high to make it shutdown
             # todo: still needed? last observed on  July 18, 2022
             self.logger.error("Temp sensor failure. Switching off. See issue #308")
             self._update_heater(0.0)
-            self.set_automation(
-                TemperatureAutomation(automation_name="only_record_ambient_temperature")
-            )
+            self.set_automation(TemperatureAutomation(automation_name="only_record_temperature"))
 
         return averaged_temp
 
@@ -286,7 +284,7 @@ class TemperatureController(BackgroundJob):
 
             # since we are changing automations inside a controller, we know that the latest temperature reading is recent, so we can
             # pass it on to the new automation.
-            # this is most useful when temp-control is initialized with only_record_ambient_temperature, and then quickly switched over to thermostat.
+            # this is most useful when temp-control is initialized with only_record_temperature, and then quickly switched over to thermostat.
             self.automation_job._set_latest_temperature(self.temperature)
 
         except KeyError:
@@ -326,14 +324,14 @@ class TemperatureController(BackgroundJob):
             self.blink_error_code(error_codes.PCB_TEMPERATURE_TOO_HIGH)
 
             self.logger.warning(
-                f"Temperature of heating surface has exceeded {self.MAX_TEMP_TO_DISABLE_HEATING}℃ - currently {temp}℃. This is beyond our recommendations. The heating PWM channel will be forced to 0 and the automation turned to only_record_ambient_temperature. Take caution when touching the heating surface and wetware."
+                f"Temperature of heating surface has exceeded {self.MAX_TEMP_TO_DISABLE_HEATING}℃ - currently {temp}℃. This is beyond our recommendations. The heating PWM channel will be forced to 0 and the automation turned to only_record_temperature. Take caution when touching the heating surface and wetware."
             )
 
             self._update_heater(0)
 
-            if self.automation_name != "only_record_ambient_temperature":
+            if self.automation_name != "only_record_temperature":
                 self.set_automation(
-                    TemperatureAutomation(automation_name="only_record_ambient_temperature")
+                    TemperatureAutomation(automation_name="only_record_temperature")
                 )
 
         elif temp > self.MAX_TEMP_TO_REDUCE_HEATING:
@@ -608,7 +606,7 @@ def start_temperature_control(
 )
 @click.option(
     "--automation-name",
-    default="only_record_ambient_temperature",
+    default="only_record_temperature",
     help="set the automation of the system",
     show_default=True,
 )